Functions for options in a workflow set

Functions for options in a workflow set

add_options(x, ..., strict = FALSE)

remove_options(x, ...)

add_option_parameters(x, strict = FALSE)

Arguments

x
A workflow set

...
A list of named options. For remove_options() this can be a series of unquoted option names.

strict
A logical; show execution stop if existing options are being replaced?

Value

An updated workflow set.

Details

add_options() is used to update all of the options in a workflow set.

remove_options() will eliminate specific options across rows.

add_option_parameters() adds a parameter object to the option column (if parameters are being tuned).

Note that executing a function on the workflow set, such as tune_grid(), will add any options given to that funciton to the option column.