Obtain and format results produced by tuning functions for workflow sets

Obtain and format results produced by tuning functions for workflow sets

# S3 method for workflow_set
collect_metrics(x, summarize = TRUE, ...)

# S3 method for workflow_set
collect_predictions(
  x,
  summarize = TRUE,
  parameters = NULL,
  select_best = FALSE,
  metric = NULL,
  ...
)

Arguments

x
A workflow_set object where all workflows have been evaluated.

summarize
A logical for whether the performance estimates should be summarized via the mean (over resamples) or the raw performance values (per resample) should be returned along with the resampling identifiers. When collecting predictions, these are averaged if multiple assessment sets contain the same row.

...
Not currently used.

parameters
An optional tibble of tuning parameter values that can be used to filter the predicted values before processing. This tibble should only have columns for each tuning parameter identifier (e.g. "my_param" if tune("my_param") was used).

select_best
A single logical for whether the numerically best results are retained. If TRUE, the parameters argument is ignored.

metric
A character string for the metric that is used for select_best.

Value

A tibble.

Details

When applied to a workflow set, the metrics and predictions that are returned do not contain the actual tuning parameter columns and values (unlike when these collect functions are run on other objects). The reason is that workflow sets can contain different types of models or models with different tuning parameters.

If the columns are needed, there are two options. First, the .config column can be used to merge the tuning parameter columns into an appropriate object. Alternatively, the map() function can be used to get the metrics from the original objects (see the example below).

Examples

library(workflowsets)
library(modeldata)
library(recipes)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> 
#> Attaching package: ‘recipes’

#> The following object is masked from ‘package:stats’:
#> 
#>     step

library(parsnip)
library(dplyr)
library(rsample)
library(tune)
library(yardstick)

#> For binary classification, the first factor level is assumed to be the event.
#> Use the argument `event_level = "second"` to alter this as needed.

library(purrr)
library(tidyr)

# ------------------------------------------------------------------------------

data(cells)
cells <- cells %>% dplyr::select(-case)

set.seed(1)
val_set <- validation_split(cells)

# ------------------------------------------------------------------------------

basic_recipe <-
   recipe(class ~ ., data = cells) %>%
   step_YeoJohnson(all_predictors()) %>%
   step_normalize(all_predictors())

pca_recipe <-
   basic_recipe %>%
   step_pca(all_predictors(), num_comp = tune())

ss_recipe <-
   basic_recipe %>%
   step_spatialsign(all_predictors())

# ------------------------------------------------------------------------------

knn_mod <-
   nearest_neighbor(neighbors = tune(), weight_func = tune()) %>%
   set_engine("kknn") %>%
   set_mode("classification")

lr_mod <-
   logistic_reg() %>%
   set_engine("glm")

# ------------------------------------------------------------------------------

preproc <- list(none = basic_recipe, pca = pca_recipe, sp_sign = ss_recipe)
models <- list(knn = knn_mod, logistic = lr_mod)

cell_models <- workflow_set(preproc, models, cross = TRUE)

# ------------------------------------------------------------------------------

# \donttest{
cell_models <-
   cell_models %>%
   workflow_map(resamples = val_set, grid = 10, metrics = metric_set(roc_auc))

#> ℹ No tuning parameters. `fit_resamples()` will be attempted

#> ℹ No tuning parameters. `fit_resamples()` will be attempted

cell_models

#> # A workflow set/tibble: 6 x 6
#>   wflow_id        preproc model          workflow  option         result        
#>   <chr>           <chr>   <chr>          <list>    <list>         <list>        
#> 1 none_knn        recipe  nearest_neigh… <workflo… <named list [… <tibble [1 × …
#> 2 none_logistic   recipe  logistic_reg   <workflo… <named list [… <tibble [1 × …
#> 3 pca_knn         recipe  nearest_neigh… <workflo… <named list [… <tibble [1 × …
#> 4 pca_logistic    recipe  logistic_reg   <workflo… <named list [… <tibble [1 × …
#> 5 sp_sign_knn     recipe  nearest_neigh… <workflo… <named list [… <tibble [1 × …
#> 6 sp_sign_logist… recipe  logistic_reg   <workflo… <named list [… <tibble [1 × …

collect_metrics(cell_models)

#> # A tibble: 36 x 9
#>    wflow_id .config      preproc model    .metric .estimator  mean     n std_err
#>    <chr>    <chr>        <chr>   <chr>    <chr>   <chr>      <dbl> <int>   <dbl>
#>  1 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.873     1      NA
#>  2 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.866     1      NA
#>  3 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.888     1      NA
#>  4 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.876     1      NA
#>  5 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.891     1      NA
#>  6 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.848     1      NA
#>  7 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.886     1      NA
#>  8 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.887     1      NA
#>  9 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.870     1      NA
#> 10 none_knn Preprocesso… recipe  nearest… roc_auc binary     0.772     1      NA
#> # … with 26 more rows

# Alternatively, if the tuning parameter values are needed:
cell_models %>%
  dplyr::filter(grepl("knn", wflow_id)) %>%
  mutate(metrics = map(result, collect_metrics)) %>%
  dplyr::select(wflow_id, metrics) %>%
  tidyr::unnest(cols = metrics)

#> # A tibble: 30 x 10
#>    wflow_id neighbors weight_func .metric .estimator  mean     n std_err .config
#>    <chr>        <int> <chr>       <chr>   <chr>      <dbl> <int>   <dbl> <chr>  
#>  1 none_knn         9 biweight    roc_auc binary     0.873     1      NA Prepro…
#>  2 none_knn         5 cos         roc_auc binary     0.866     1      NA Prepro…
#>  3 none_knn        13 epanechnik… roc_auc binary     0.888     1      NA Prepro…
#>  4 none_knn         7 gaussian    roc_auc binary     0.876     1      NA Prepro…
#>  5 none_knn        14 inv         roc_auc binary     0.891     1      NA Prepro…
#>  6 none_knn         3 optimal     roc_auc binary     0.848     1      NA Prepro…
#>  7 none_knn        11 rank        roc_auc binary     0.886     1      NA Prepro…
#>  8 none_knn        10 rectangular roc_auc binary     0.887     1      NA Prepro…
#>  9 none_knn         6 triangular  roc_auc binary     0.870     1      NA Prepro…
#> 10 none_knn         1 triweight   roc_auc binary     0.772     1      NA Prepro…
#> # … with 20 more rows, and 1 more variable: num_comp <int>

# }