Methods for processing workflow sets

Methods for processing workflow sets

# S3 method for workflow_set
tune_grid(object, ..., which = NULL, seed = sample(1e+05, 1), verbose = FALSE)

# S3 method for workflow_set
tune_bayes(object, ..., which = NULL, seed = sample(1e+05, 1), verbose = FALSE)

# S3 method for workflow_set
fit_resamples(
  object,
  ...,
  which = NULL,
  seed = sample(1e+05, 1),
  verbose = FALSE
)

Arguments

object
A workflow_set object.

...
Common arguments that will be passed to each tune function. These should not be the same arguments passed to options.

which
A character string of identifiers indicating which workflows should be processed. If left NULL, all applicable workflows without entries in the results column are selected.

seed
A single integer that is set before each workflow is processed.

verbose
A single logical to indicate if logging should occur. This will write a message after each workflow is process and indicates if there were any issues.

Value

An object from the corresponding tune package function (e.g. with class tune_result, etc.)

Examples

library(workflowsets)
library(modeldata)
library(recipes)
library(parsnip)
library(dplyr)
library(rsample)
library(tune)
library(yardstick)
library(dials)

# ------------------------------------------------------------------------------

data(cells)
cells <- cells %>% dplyr::select(-case)

set.seed(1)
val_set <- validation_split(cells)

# ------------------------------------------------------------------------------

basic_recipe <-
   recipe(class ~ ., data = cells) %>%
   step_YeoJohnson(all_predictors()) %>%
   step_normalize(all_predictors())

pca_recipe <-
   basic_recipe %>%
   step_pca(all_predictors(), num_comp = tune())

ss_recipe <-
   basic_recipe %>%
   step_spatialsign(all_predictors())

# ------------------------------------------------------------------------------

knn_mod <-
   nearest_neighbor(neighbors = tune(), weight_func = tune()) %>%
   set_engine("kknn") %>%
   set_mode("classification")

lr_mod <-
   logistic_reg() %>%
   set_engine("glm")

# ------------------------------------------------------------------------------

preproc <- list(none = basic_recipe, pca = pca_recipe, sp_sign = ss_recipe)
models <- list(knn = knn_mod, logistic = lr_mod)

cell_models <- workflow_set(preproc, models, cross = TRUE)

# ------------------------------------------------------------------------------

# \donttest{
cell_model_results <-
   cell_models %>%
   tune_grid(resamples = val_set, grid = 10, metrics = metric_set(roc_auc)) %>%
   fit_resamples(resamples = val_set, metrics = metric_set(roc_auc))

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

#> Warning: Unknown or uninitialised column: `options`.

cell_model_results

#> # A workflow set/tibble: 6 x 6
#>   wflow_id         preproc model            object     option     result        
#>   <chr>            <chr>   <chr>            <list>     <list>     <list>        
#> 1 none_knn         recipe  nearest_neighbor <workflow> <list [0]> <try-errr [1]>
#> 2 none_logistic    recipe  logistic_reg     <workflow> <list [0]> <try-errr [1]>
#> 3 pca_knn          recipe  nearest_neighbor <workflow> <list [0]> <try-errr [1]>
#> 4 pca_logistic     recipe  logistic_reg     <workflow> <list [0]> <try-errr [1]>
#> 5 sp_sign_knn      recipe  nearest_neighbor <workflow> <list [0]> <try-errr [1]>
#> 6 sp_sign_logistic recipe  logistic_reg     <workflow> <list [0]> <try-errr [1]>

# }

# ------------------------------------------------------------------------------

# An example of setting options. Let's change the range for `num_comp` by
# passing a specific parameter set.
# pca_knn_mod <-
#   cell_models %>%
#   dplyr::filter(wflow_id == "pca_knn")
# pca_knn_param <-
#   pca_knn_mod$object[[1]] %>%
#   parameters() %>%
#   update(num_comp = num_comp(c(0, 20)))
#
# new_opts <- list(pca_knn = list(param_info = pca_knn_param))
# new_opts
# \donttest{
# cell_model_results <-
#    cell_models %>%
#    tune_grid(resamples = val_set, grid = 10, which = "pca_knn", options = new_opts)
# }