Example Data Sets

Example Data Sets

Value

Workflow sets.

Details

Example workflow sets and associated model fits.

two_class_set and two_class_res were generated using the data in the package file example-data/two-class-set.R

chi_features_set and chi_features_res were generated using the data in the package file example-data/chi-features-res.R. It is meant to approximate the sequence of models built in Section 1.3 of Kuhn and Johnson (2019).

References

Max Kuhn and Kjell Johnson (2019) Feature Engineering and Selection, https://bookdown.org/max/FES/a-more-complex-example.html

Examples

data(two_class_set)
two_class_set
#> # A workflow set/tibble: 6 × 4
#>   wflow_id      info             option    result    
#>   <chr>         <list>           <list>    <list>    
#> 1 none_cart     <tibble [1 × 4]> <opts[0]> <list [0]>
#> 2 none_glm      <tibble [1 × 4]> <opts[0]> <list [0]>
#> 3 none_mars     <tibble [1 × 4]> <opts[0]> <list [0]>
#> 4 yj_trans_cart <tibble [1 × 4]> <opts[0]> <list [0]>
#> 5 yj_trans_glm  <tibble [1 × 4]> <opts[0]> <list [0]>
#> 6 yj_trans_mars <tibble [1 × 4]> <opts[0]> <list [0]>