Process a series of workflows

workflow_map() will execute the same function across the workflows in the set. The various tune_*() functions can be used as well as fit_resamples().

workflow_map(
  object,
  fn = "tune_grid",
  verbose = FALSE,
  seed = sample.int(10^4, 1),
  ...
)

Arguments

object
A workflow set.

fn
The function to run. Acceptable values are: tune::tune_grid(), tune::tune_bayes(), tune::fit_resamples(), finetune::tune_race_anova(), finetune::tune_race_win_loss(), or finetune::tune_sim_anneal().

verbose
A logical for logging progress.

seed
A single integer that is set prior to each function execution.

...
Options to pass to the modeling function. See details below.

Value

An updated workflow set.

Details

When passing options, anything passed in the ... will be combined with any values in the option column. The values in ... will override that column's values and the new options are added to the options column.

Any failures in execution result in the corresponding row of results to contain a try-error object.

Examples

# Duplicating the existing results
if (FALSE) {
   two_class_res <-
      two_class_set %>%
      workflow_map(resamples = folds, grid = 10, seed = 2, verbose = TRUE)
}

rank_results(chi_features_res)

#> # A tibble: 40 x 9
#>    wflow_id   .config      .metric  mean std_err     n preprocessor model   rank
#>    <chr>      <chr>        <chr>   <dbl>   <dbl> <int> <chr>        <chr>  <int>
#>  1 plus_pca_… Preprocesso… rmse    0.574      NA     1 recipe       linea…     1
#>  2 plus_pca_… Preprocesso… rsq     0.989      NA     1 recipe       linea…     1
#>  3 plus_pca_… Preprocesso… rmse    0.586      NA     1 recipe       linea…     2
#>  4 plus_pca_… Preprocesso… rsq     0.989      NA     1 recipe       linea…     2
#>  5 plus_pca_… Preprocesso… rmse    0.590      NA     1 recipe       linea…     3
#>  6 plus_pca_… Preprocesso… rsq     0.988      NA     1 recipe       linea…     3
#>  7 plus_pca_… Preprocesso… rmse    0.591      NA     1 recipe       linea…     4
#>  8 plus_pca_… Preprocesso… rsq     0.988      NA     1 recipe       linea…     4
#>  9 plus_pca_… Preprocesso… rmse    0.594      NA     1 recipe       linea…     5
#> 10 plus_pca_… Preprocesso… rsq     0.989      NA     1 recipe       linea…     5
#> # … with 30 more rows