Process a series of workflows

workflow_map() will execute the same function across the workflows in the set. The various tune_*() functions can be used as well as tune::fit_resamples().

workflow_map(
  object,
  fn = "tune_grid",
  verbose = FALSE,
  seed = sample.int(10^4, 1),
  ...
)

Arguments

object
A workflow set.

fn
The function to run. Acceptable values are: tune::tune_grid(), tune::tune_bayes(), tune::fit_resamples(), finetune::tune_race_anova(), finetune::tune_race_win_loss(), or finetune::tune_sim_anneal().

verbose
A logical for logging progress.

seed
A single integer that is set prior to each function execution.

...
Options to pass to the modeling function. See details below.

Value

An updated workflow set. The option column will be updated with any options for the tune package functions given to workflow_map(). Also, the results will be added to the result column. If the computations for a workflow fail, an try-catch object will be saved in place of the results (without stopping execution).

Details

When passing options, anything passed in the ... will be combined with any values in the option column. The values in ... will override that column's values and the new options are added to the options column.

Any failures in execution result in the corresponding row of results to contain a try-error object.

In cases where a model has no tuning parameters is mapped to one of the tuning functions, tune::fit_resamples() will be used instead and a warning is issued if verbose = TRUE.

If a workflow required packages that are not installed, a message is printed and workflow_map() continues with the next workflow (if any).

See also

Examples

# An example of processed results
chi_features_res

#> # A workflow set/tibble: 3 x 4
#>   wflow_id         info             option    result   
#>   <chr>            <list>           <list>    <list>   
#> 1 date_lm          <tibble [1 × 4]> <opts[2]> <rsmp[+]>
#> 2 plus_holidays_lm <tibble [1 × 4]> <opts[2]> <rsmp[+]>
#> 3 plus_pca_lm      <tibble [1 × 4]> <opts[3]> <tune[+]>

# Code examples at
if (interactive()) {
  system.file("example-data", package = "workflowsets")
}