Process a series of workflows

workflow_map() will execute the same function across the workflows in the set. The various tune_*() functions can be used as well as fit_resamples().

workflow_map(
  object,
  fn = "tune_grid",
  verbose = FALSE,
  seed = sample.int(10^4, 1),
  ...
)

Arguments

object
A workflow set.

fn
The function to run. Acceptable values are: tune::tune_grid(), tune::tune_bayes(), tune::fit_resamples(), finetune::tune_race_anova(), finetune::tune_race_win_loss(), or finetune::tune_sim_anneal().

verbose
A logical for logging progress.

seed
A single integer that is set prior to each function execution.

...
Options to pass to the modeling function. See details below.

Value

An updated workflow set.

Details

When passing options, anything passed in the ... will be combined with any values in the option column. The values in ... will override that column's values and the new options are added to the options column.

Any failures in execution result in the corresponding row of results to contain a try-error object.

Examples

example("workflow_set", run.donttest = FALSE)

#> 
#> wrkfl_> library(workflowsets)
#> 
#> wrkfl_> library(modeldata)
#> 
#> wrkfl_> library(recipes)
#> 
#> wrkfl_> library(parsnip)
#> 
#> wrkfl_> library(dplyr)
#> 
#> wrkfl_> library(rsample)
#> 
#> wrkfl_> library(tune)
#> 
#> wrkfl_> library(yardstick)
#> 
#> wrkfl_> # ------------------------------------------------------------------------------
#> wrkfl_> 
#> wrkfl_> data(cells)
#> 
#> wrkfl_> cells <- cells %>% dplyr::select(-case)
#> 
#> wrkfl_> set.seed(1)
#> 
#> wrkfl_> val_set <- validation_split(cells)
#> 
#> wrkfl_> # ------------------------------------------------------------------------------
#> wrkfl_> 
#> wrkfl_> basic_recipe <-
#> wrkfl_+    recipe(class ~ ., data = cells) %>%
#> wrkfl_+    step_YeoJohnson(all_predictors()) %>%
#> wrkfl_+    step_normalize(all_predictors())
#> 
#> wrkfl_> pca_recipe <-
#> wrkfl_+    basic_recipe %>%
#> wrkfl_+    step_pca(all_predictors(), num_comp = tune())
#> 
#> wrkfl_> ss_recipe <-
#> wrkfl_+    basic_recipe %>%
#> wrkfl_+    step_spatialsign(all_predictors())
#> 
#> wrkfl_> # ------------------------------------------------------------------------------
#> wrkfl_> 
#> wrkfl_> knn_mod <-
#> wrkfl_+    nearest_neighbor(neighbors = tune(), weight_func = tune()) %>%
#> wrkfl_+    set_engine("kknn") %>%
#> wrkfl_+    set_mode("classification")
#> 
#> wrkfl_> lr_mod <-
#> wrkfl_+    logistic_reg() %>%
#> wrkfl_+    set_engine("glm")
#> 
#> wrkfl_> # ------------------------------------------------------------------------------
#> wrkfl_> 
#> wrkfl_> preproc <- list(none = basic_recipe, pca = pca_recipe, sp_sign = ss_recipe)
#> 
#> wrkfl_> models <- list(knn = knn_mod, logistic = lr_mod)
#> 
#> wrkfl_> cell_models <- workflow_set(preproc, models, cross = TRUE)
#> 
#> wrkfl_> cell_models
#> # A workflow set/tibble: 6 x 6
#>   wflow_id         preproc model            workflow   option     result    
#>   <chr>            <chr>   <chr>            <list>     <list>     <list>    
#> 1 none_knn         recipe  nearest_neighbor <workflow> <list [0]> <list [0]>
#> 2 none_logistic    recipe  logistic_reg     <workflow> <list [0]> <list [0]>
#> 3 pca_knn          recipe  nearest_neighbor <workflow> <list [0]> <list [0]>
#> 4 pca_logistic     recipe  logistic_reg     <workflow> <list [0]> <list [0]>
#> 5 sp_sign_knn      recipe  nearest_neighbor <workflow> <list [0]> <list [0]>
#> 6 sp_sign_logistic recipe  logistic_reg     <workflow> <list [0]> <list [0]>
#> 
#> wrkfl_> # ------------------------------------------------------------------------------
#> wrkfl_> # Using selector functions and formulas
#> wrkfl_> 
#> wrkfl_> # Select predictors by their names
#> wrkfl_> channels <- paste0("ch_", 1:4)
#> 
#> wrkfl_> preproc <- purrr::map(channels, ~ selectors(class, c(contains(!!.x))))
#> 
#> wrkfl_> names(preproc) <- channels
#> 
#> wrkfl_> preproc$everything <- class ~ .
#> 
#> wrkfl_> preproc
#> $ch_1
#> A set of tidy selectors for model columns.
#> 
#> $ch_2
#> A set of tidy selectors for model columns.
#> 
#> $ch_3
#> A set of tidy selectors for model columns.
#> 
#> $ch_4
#> A set of tidy selectors for model columns.
#> 
#> $everything
#> class ~ .
#> 
#> 
#> wrkfl_> cell_models_by_group <- workflow_set(preproc, models["logistic"], cross = TRUE)

cell_fits <-
   cell_models_by_group %>%
   workflow_map("fit_resamples",
                resamples = val_set,
                metrics = metric_set(roc_auc))

#> ! validation: preprocessor 1/1, model 1/1 (predictions): prediction from a rank-defici...

#> ! validation: preprocessor 1/1, model 1/1 (predictions): prediction from a rank-defici...

rank_results(cell_fits)

#> # A tibble: 5 x 9
#>   wflow_id     .config    .metric  mean std_err     n model   preprocessor  rank
#>   <chr>        <chr>      <chr>   <dbl>   <dbl> <int> <chr>   <chr>        <int>
#> 1 everything_… Preproces… roc_auc 0.863      NA     1 logist… formula          1
#> 2 ch_1_logist… Preproces… roc_auc 0.859      NA     1 logist… selector         2
#> 3 ch_2_logist… Preproces… roc_auc 0.808      NA     1 logist… selector         3
#> 4 ch_3_logist… Preproces… roc_auc 0.768      NA     1 logist… selector         4
#> 5 ch_4_logist… Preproces… roc_auc 0.742      NA     1 logist… selector         5