Generate a set of workflow objects from preprocessing and model objects

Generate a set of workflow objects from preprocessing and model objects

workflow_set(preproc, models, cross = TRUE)

Arguments

preproc
A list (preferably named) with preprocessing objects: formulas, recipes, or selectors.

models
A list (preferably named) of parsnip model specifications.

cross
A logical: should all combinations of the preprocessors and models be used to create the workflows? If FALSE, the length of preproc and models should be equal.

Value

A tibble with extra class 'workflow_set'.

Examples

library(workflowsets)
library(modeldata)
library(recipes)
library(parsnip)
library(dplyr)
library(rsample)
library(tune)
library(yardstick)

# ------------------------------------------------------------------------------

data(cells)
cells <- cells %>% dplyr::select(-case)

set.seed(1)
val_set <- validation_split(cells)

# ------------------------------------------------------------------------------

basic_recipe <-
   recipe(class ~ ., data = cells) %>%
   step_YeoJohnson(all_predictors()) %>%
   step_normalize(all_predictors())

pca_recipe <-
   basic_recipe %>%
   step_pca(all_predictors(), num_comp = tune())

ss_recipe <-
   basic_recipe %>%
   step_spatialsign(all_predictors())

# ------------------------------------------------------------------------------

knn_mod <-
   nearest_neighbor(neighbors = tune(), weight_func = tune()) %>%
   set_engine("kknn") %>%
   set_mode("classification")

lr_mod <-
   logistic_reg() %>%
   set_engine("glm")

# ------------------------------------------------------------------------------

preproc <- list(none = basic_recipe, pca = pca_recipe, sp_sign = ss_recipe)
models <- list(knn = knn_mod, logistic = lr_mod)

cell_models <- workflow_set(preproc, models, cross = TRUE)
cell_models

#> # A workflow set/tibble: 6 x 4
#>   wflow_id         info             option     result    
#>   <chr>            <list>           <list>     <list>    
#> 1 none_knn         <tibble [1 × 3]> <list [0]> <list [0]>
#> 2 none_logistic    <tibble [1 × 3]> <list [0]> <list [0]>
#> 3 pca_knn          <tibble [1 × 3]> <list [0]> <list [0]>
#> 4 pca_logistic     <tibble [1 × 3]> <list [0]> <list [0]>
#> 5 sp_sign_knn      <tibble [1 × 3]> <list [0]> <list [0]>
#> 6 sp_sign_logistic <tibble [1 × 3]> <list [0]> <list [0]>

# ------------------------------------------------------------------------------
# Using selector functions and formulas

# Select predictors by their names
channels <- paste0("ch_", 1:4)
preproc <- purrr::map(channels, ~ selectors(class, c(contains(!!.x))))
names(preproc) <- channels
preproc$everything <- class ~ .
preproc

#> $ch_1
#> A set of tidy selectors for model columns.
#> 
#> $ch_2
#> A set of tidy selectors for model columns.
#> 
#> $ch_3
#> A set of tidy selectors for model columns.
#> 
#> $ch_4
#> A set of tidy selectors for model columns.
#> 
#> $everything
#> class ~ .
#> <environment: 0x7fdc94e84b90>
#> 

cell_models_by_group <- workflow_set(preproc, models["logistic"], cross = TRUE)